+++ Index.php
@@ -18,20 +18,25 @@
         <?php
             require_once 'ClassNote.php';
 
+            $Submitted = false;
+            if(!empty($_POST["SubmitPressed"])) {
+                $Submitted = $_POST["SubmitPressed"];
+            }
+
             session_start();
-            if(!empty($_SESSION["NoteToDelete"])) {
+            if(!empty($_SESSION["NoteToDelete"]) && $Submitted) {
 
                 $NoteToDelete = $_SESSION["NoteToDelete"];
                 $NoteToDelete->DeleteDataset();
             }
             session_abort();
 
-            if(!empty($_POST["Title"]) && !empty($_POST["Author"]) && !empty($_POST["Note"])) {
+            if(!empty($_POST["Title"]) && !empty($_POST["Author"]) && !empty($_POST["NewNote"])) {
                 $Title = htmlspecialchars(trim($_POST["Title"]));
                 $Author = htmlspecialchars(trim($_POST["Author"]));
-                $Note = htmlspecialchars(trim($_POST["Note"]));
+                $Note = htmlspecialchars(trim($_POST["NewNote"]));
 
-                $NoteObject = new Note($Title, $Author, $Notes);
+                $NoteObject = new Note($Title, $Author, $Note);
                 $NoteObject->InsertDataset();
             }
         ?>
